requirements.txt
python-dotenv==0.20.0
loguru==0.7.0
fastapi==0.100.0
dnspython==2.2.1
pymongo==4.3.3
pycryptodomex==3.18.0
